refactor front: move DropdownMenu class to separate file
